refactor(backend): Refactor and restructure backend code
what's changed:
- Moved every api component to `app/` directory
- Refactor and Restructure code
- Add documentation and annotation for files, classes and methods
